Rename directories to kebab-case
