+++ apps/src/tests/single-feature-tests/bottom-tabs/TabBarHidden.tsx
